Removed warnings

+++ WC397/4.cpp
@@ -5,10 +5,10 @@ class Solution {
 public:
     int maximumEnergy(vector<int>& energy, int k) {
 				int max_total_energy = 0;
-				for (int i = 1; i < energy.size(); i++) {
+				for (int i = 1; i < (int)energy.size(); i++) {
 					// each case: start from i
 					int total_energy = energy[i];
-					while(i+k < energy.size()) {
+					while(i+k < (int)energy.size()) {
 						total_energy += energy[i+k];
 						i += k;
 					}
